+++ src/cli/wallet.py
@@ -1,4 +1,4 @@
-from os import path, remove, environ
+from os import path, remove
 import pexpect
 import json
 import requests
@@ -6,27 +6,20 @@ import requests
 from json.decoder import JSONDecodeError 
 
 from conf.meile_config import MeileGuiConfig
-from cli.sentinel import IBCATOM, IBCDEC, IBCOSMO, IBCSCRT, SATOSHI, APIURL
-
-USER = environ['SUDO_USER'] if 'SUDO_USER' in environ else environ['USER']
-PATH = environ['PATH']
-KEYRINGDIR = path.join(path.expanduser('~' + USER), '.meile-gui')
-BASEDIR  = path.join(path.expanduser('~' + USER), '.sentinelcli')
-WALLETINFO = path.join(KEYRINGDIR, "infos.txt")
-SUBSCRIBEINFO = path.join(KEYRINGDIR, "subscribe.infos")
-CONNECTIONINFO = path.join(KEYRINGDIR, "connection.infos")
+from typedef.konstants import IBCTokens 
+from typedef.konstants import ConfParams 
+from typedef.konstants import HTTParams 
 
 MeileConfig = MeileGuiConfig()
 sentinelcli = MeileConfig.resource_path("../bin/sentinelcli")
-RPC = "https://rpc.mathnodes.com:443"
 
 class HandleWalletFunctions():
     
     def create(self, wallet_name, keyring_passphrase, seed_phrase):
-        SCMD = '%s keys add "%s" -i --keyring-backend file --keyring-dir %s' % (sentinelcli, wallet_name, KEYRINGDIR)
+        SCMD = '%s keys add "%s" -i --keyring-backend file --keyring-dir %s' % (sentinelcli, wallet_name, ConfParams.KEYRINGDIR)
         DUPWALLET = False 
         line_numbers = [11, 21]
-        ofile =  open(WALLETINFO, "wb")    
+        ofile =  open(ConfParams.WALLETINFO, "wb")    
         
         ''' Process to handle wallet in sentinel-cli '''
         child = pexpect.spawn(SCMD)
@@ -58,7 +51,7 @@ class HandleWalletFunctions():
                 child.expect(pexpect.EOF)
                 ofile.flush()
                 ofile.closae()
-                remove(WALLETINFO)
+                remove(ConfParams.WALLETINFO)
                 return None
             else:
                 child.expect(pexpect.EOF)
@@ -74,7 +67,7 @@ class HandleWalletFunctions():
         
      
         if not DUPWALLET:
-            with open(WALLETINFO, "r") as dvpn_file:
+            with open(ConfParams.WALLETINFO, "r") as dvpn_file:
                 WalletDict = {}   
                 lines = dvpn_file.readlines()
                 addy_seed = [lines[x] for x in range(line_numbers[0], line_numbers[1] +1)]
@@ -83,25 +76,25 @@ class HandleWalletFunctions():
                 else:
                     WalletDict['address'] = addy_seed[1].split(":")[-1].lstrip().rstrip()
                 WalletDict['seed'] = lines[-1].lstrip().rstrip().replace('\n', '')
-                #remove(WALLETINFO)
+                #remove(ConfParams.WALLETINFO)
                 return WalletDict
     
         else:
-            remove(WALLETINFO)
+            remove(ConfParams.WALLETINFO)
             return None
 
     
     
     def subscribe(self, KEYNAME, NODE, DEPOSIT):
-        CONFIG = MeileGuiConfig.read_configuration(MeileGuiConfig, MeileGuiConfig.CONFFILE)
+        CONFIG = MeileConfig.read_configuration(MeileConfig.CONFFILE)
         PASSWORD = CONFIG['wallet'].get('password', '')
     
-        ofile =  open(SUBSCRIBEINFO, "wb")
+        ofile =  open(ConfParams.SUBSCRIBEINFO, "wb")
             
         if not KEYNAME:
             return (False, 1337)
         
-        SCMD = "%s tx subscription subscribe-to-node --yes --keyring-backend file --keyring-dir %s --gas-prices 0.1udvpn --chain-id sentinelhub-2 --node %s --from '%s' '%s' %s"  % (sentinelcli, KEYRINGDIR, RPC, KEYNAME, NODE, DEPOSIT)    
+        SCMD = "%s tx subscription subscribe-to-node --yes --keyring-backend file --keyring-dir %s --gas-prices 0.1udvpn --chain-id sentinelhub-2 --node %s --from '%s' '%s' %s"  % (sentinelcli, ConfParams.KEYRINGDIR, HTTParams.RPC, KEYNAME, NODE, DEPOSIT)    
         try:
             child = pexpect.spawn(SCMD)
             child.logfile = ofile
@@ -120,8 +113,7 @@ class HandleWalletFunctions():
         
             
     def ParseSubscribe(self):
-        SUBSCRIBEINFO = path.join(KEYRINGDIR, "subscribe.infos")
-        with open(SUBSCRIBEINFO, 'r') as sub_file:
+        with open(ConfParams.SUBSCRIBEINFO, 'r') as sub_file:
                 lines = sub_file.readlines()
                 try:
                     tx_json = json.loads(lines[2])
@@ -135,25 +127,25 @@ class HandleWalletFunctions():
                     try: 
                         sub_id = tx_json['logs'][0]['events'][4]['attributes'][0]['value']
                         if sub_id:
-                            remove(SUBSCRIBEINFO)
+                            remove(ConfParams.SUBSCRIBEINFO)
                             return (True,0)
                         else:
-                            remove(SUBSCRIBEINFO)
+                            remove(ConfParams.SUBSCRIBEINFO)
                             return (False,2.71828) 
                     except:
-                        remove(SUBSCRIBEINFO)
+                        remove(ConfParams.SUBSCRIBEINFO)
                         return (False, 3.14159)
                 elif 'insufficient' in tx_json['raw_log']:
-                    remove(SUBSCRIBEINFO)
+                    remove(ConfParams.SUBSCRIBEINFO)
                     return (False, tx_json['raw_log'])
     def connect(self, ID, address):
 
-        CONFIG = MeileGuiConfig.read_configuration(MeileGuiConfig, MeileGuiConfig.CONFFILE)
+        CONFIG = MeileConfig.read_configuration(MeileConfig.CONFFILE)
         PASSWORD = CONFIG['wallet'].get('password', '')
         KEYNAME = CONFIG['wallet'].get('keyname', '')
-        connCMD = "pkexec env PATH=%s %s connect --home %s --keyring-backend file --keyring-dir %s --chain-id sentinelhub-2 --node %s --gas-prices 0.1udvpn --yes --from '%s' %s %s" % (PATH, sentinelcli, BASEDIR, KEYRINGDIR, RPC, KEYNAME, ID, address)
+        connCMD = "pkexec env PATH=%s %s connect --home %s --keyring-backend file --keyring-dir %s --chain-id sentinelhub-2 --node %s --gas-prices 0.1udvpn --yes --from '%s' %s %s" % (ConfParams.PATH, sentinelcli, ConfParams.BASEDIR, ConfParams.KEYRINGDIR, HTTParams.RPC, KEYNAME, ID, address)
         
-        ofile =  open(CONNECTIONINFO, "wb")    
+        ofile =  open(ConfParams.CONNECTIONINFO, "wb")    
 
         try:
             child = pexpect.spawn(connCMD)
@@ -168,7 +160,7 @@ class HandleWalletFunctions():
         except pexpect.exceptions.TIMEOUT:
             return False
         
-        if path.isfile(path.join(BASEDIR, "status.json")):
+        if path.isfile(ConfParams.WIREGUARD_STATUS):
             CONNECTED = True
         else:
             CONNECTED = False
@@ -176,10 +168,11 @@ class HandleWalletFunctions():
         return CONNECTED
 
     def get_balance(self, address):
-        endpoint = "/bank/balances/" + address
+        endpoint = HTTParams.BALANCES_ENDPOINT + address
         CoinDict = {'dvpn' : 0, 'scrt' : 0, 'dec'  : 0, 'atom' : 0, 'osmo' : 0}
+        
         try:
-            r = requests.get(APIURL + endpoint)
+            r = requests.get(HTTParams.APIURL + endpoint, timeout=HTTParams.TIMEOUT)
             coinJSON = r.json()
         except:
             return None
@@ -188,15 +181,15 @@ class HandleWalletFunctions():
         try:
             for coin in coinJSON['result']:
                 if "udvpn" in coin['denom']:
-                    CoinDict['dvpn'] = round(float(float(coin['amount']) / SATOSHI),4)
-                elif IBCSCRT in coin['denom']:
-                    CoinDict['scrt'] = round(float(float(coin['amount']) / SATOSHI),4)
-                elif IBCDEC in coin['denom']:
-                    CoinDict['dec'] = round(float(float(coin['amount']) / SATOSHI),4)
-                elif IBCATOM in coin['denom']:
-                    CoinDict['atom'] = round(float(float(coin['amount']) / SATOSHI),4)
-                elif IBCOSMO in coin['denom']:
-                    CoinDict['osmo'] = round(float(float(coin['amount']) / SATOSHI),4)
+                    CoinDict['dvpn'] = round(float(float(coin['amount']) /IBCTokens.SATOSHI),4)
+                elif IBCTokens.IBCSCRT in coin['denom']:
+                    CoinDict['scrt'] = round(float(float(coin['amount']) /IBCTokens.SATOSHI),4)
+                elif IBCTokens.IBCDEC in coin['denom']:
+                    CoinDict['dec'] = round(float(float(coin['amount']) /IBCTokens.SATOSHI),4)
+                elif IBCTokens.IBCATOM in coin['denom']:
+                    CoinDict['atom'] = round(float(float(coin['amount']) /IBCTokens.SATOSHI),4)
+                elif IBCTokens.IBCOSMO in coin['denom']:
+                    CoinDict['osmo'] = round(float(float(coin['amount']) /IBCTokens.SATOSHI),4)
         except Exception as e:
             print(str(e))
             return None